move fixtures DBAccessorTest to own file
